pyproject.toml
[tool.poetry]
name = "tg-bot-for-habit-tracking"
version = "0.1.0"
description = ""
authors = ["Artyom Globin <superartem2012@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.4"
bcrypt = "^4.1.3"
sqlalchemy = "^2.0.31"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = "^2.8.2"
fastapi = "^0.111.0"
pytelegrambotapi = "^4.21.0"
pydantic-settings = "^2.3.4"
uvicorn = "^0.30.1"
asyncpg = "^0.29.0"
aiohttp = "^3.9.5"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"
apscheduler = "^3.10.4"
black = "^24.8.0"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
